refactor(types,ui): move style types from ui to types package
Extract NubeComponentStyle and theme types to eliminate circular dependencies.
Move csstype dependency to types package where it's actually used.
