+++ packages/types/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@tiendanube/nube-sdk-types",
-	"version": "0.25.0",
+	"version": "0.26.0",
 	"description": "Type definition of NubeSDK",
 	"type": "module",
 	"main": "./src/index.ts",
@@ -19,12 +19,12 @@
 		"url": "https://github.com/TiendaNube/nube-sdk/issues"
 	},
 	"homepage": "https://www.tiendanube.com",
+	"dependencies": {
+		"csstype": "^3.1.3"
+	},
 	"devDependencies": {
 		"@biomejs/biome": "1.8.3",
 		"typescript": "^5.6.2"
-	},
-	"peerDependencies": {
-		"@tiendanube/nube-sdk-ui": "*"
 	},
 	"files": ["./src", "README.md", "CHANGELOG.md"],
 	"keywords": ["tiendanube", "nuvemshop", "nube-sdk", "ecommerce", "sdk"]
