+++ packages/ui/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@tiendanube/nube-sdk-ui",
-  "version": "0.10.0",
+  "version": "0.11.0",
   "description": "Library for building declarative interfaces for NubeSDK",
   "type": "module",
   "main": "./dist/index.js",
@@ -26,7 +26,6 @@
   "devDependencies": {
     "@biomejs/biome": "1.8.3",
     "@vitest/coverage-v8": "^3.1.1",
-    "csstype": "^3.1.3",
     "tsup": "^8.3.5",
     "typescript": "^5.6.2",
     "vitest": "^3.1.1"
