+++ packages/ui/src/styles/StyleSheet.ts
@@ -1,51 +1,5 @@
-import type { Size } from "@tiendanube/nube-sdk-types";
-import type * as CSS from "csstype";
+import type { NubeComponentStyle } from "@tiendanube/nube-sdk-types";
 import { ThemeColor } from "./ThemeColor";
-import type { ThemeCSSValue } from "./theme";
-
-/**
- * Maps properties that should use the Size type
- */
-type SizePropertyKeys =
-	| "width"
-	| "height"
-	| "minWidth"
-	| "minHeight"
-	| "maxWidth"
-	| "maxHeight"
-	| "top"
-	| "right"
-	| "bottom"
-	| "left"
-	| "margin"
-	| "marginTop"
-	| "marginBottom"
-	| "marginLeft"
-	| "marginRight"
-	| "padding"
-	| "paddingTop"
-	| "paddingBottom"
-	| "paddingLeft"
-	| "paddingRight"
-	| "fontSize"
-	| "lineHeight"
-	| "borderWidth"
-	| "borderRadius";
-
-/**
- * Applies Size only to size properties.
- * The others remain as string | number.
- */
-type EnhancedCSSProperties = {
-	[K in keyof CSS.Properties]?: K extends SizePropertyKeys
-		? Size | ThemeCSSValue
-		: CSS.Properties[K] | ThemeCSSValue;
-};
-
-/**
- * Define named styles
- */
-export type NubeComponentStyle = Partial<EnhancedCSSProperties>;
 
 function parseValue(value: unknown): unknown {
 	if (value instanceof ThemeColor) {
